http://www.atutor.ca/atutor/mantis/view.php?id=3737

+++ docs/include/html/enroll_edit.inc.php
@@ -107,34 +107,37 @@ function unenroll ($list) {
 		$sql = "SELECT forum_id from ".TABLE_PREFIX."forums_courses WHERE course_id = '$this_course_id'";
 		$result = mysql_query($sql, $db);
 
-		while($row = mysql_fetch_assoc($result)){
-			$this_course_forums[] = $row['forum_id'];
-		}
-		$this_forum_list = implode(',', $this_course_forums);
+		if($result && mysql_num_rows($result)>0){
+			while($row = mysql_fetch_assoc($result)){
+				$this_course_forums[] = $row['forum_id'];
+			}
+			$this_forum_list = implode(',', $this_course_forums);
 
-		// delete from forum_subscription any member in $members (being unenrolled)
-		// with posts to forums in this course. 
-		foreach ($this_course_forums as $this_course_forum){
-			$sql1 = "DELETE FROM ".TABLE_PREFIX."forums_subscriptions WHERE forum_id = '$this_course_forum' AND member_id IN ($members)";
-			$result_unsub = mysql_query($sql1, $db);
+			// delete from forum_subscription any member in $members (being unenrolled)
+			// with posts to forums in this course. 
+			foreach ($this_course_forums as $this_course_forum){
+				$sql1 = "DELETE FROM ".TABLE_PREFIX."forums_subscriptions WHERE forum_id = '$this_course_forum' AND member_id IN ($members)";
+				$result_unsub = mysql_query($sql1, $db);
+			}
 		}
 
 		// get a list of posts for forums in the current course
 		$sql = "SELECT post_id FROM ".TABLE_PREFIX."forums_threads WHERE forum_id IN ($this_forum_list)";
 		$result = mysql_query($sql, $db);
-		while($row = mysql_fetch_assoc($result)){
-			$this_course_posts[] = $row['post_id'];
+		if($result && mysql_num_rows($result)>0){
+			while($row = mysql_fetch_assoc($result)){
+				$this_course_posts[] = $row['post_id'];
+			}
+			$this_post_list = implode(',', $this_course_posts);
+
+			// delete from forums_accessed any post with member_id in $members being unenrolled, 
+			// and post_id in 
+			foreach($this_course_posts as $this_course_post){
+
+				$sql2	= "DELETE FROM ".TABLE_PREFIX."forums_accessed WHERE post_id = '$this_course_post' AND member_id IN ($members)";
+				$result_unsub2 = mysql_query($sql2, $db);
+			}
 		}
-		$this_post_list = implode(',', $this_course_posts);
-
-		// delete from forums_accessed any post with member_id in $members being unenrolled, 
-		// and post_id in 
-		foreach($this_course_posts as $this_course_post){
-
-			$sql2	= "DELETE FROM ".TABLE_PREFIX."forums_accessed WHERE post_id = '$this_course_post' AND member_id IN ($members)";
-			$result_unsub2 = mysql_query($sql2, $db);
-		}
-
 	}
 }
